cosas

+++ web/static/js/perdido.js
@@ -24,37 +24,44 @@ function getDatosPerdidos()
 
 function deletePerdido(id)
 {
-	//obtiene el csrftoken
-	var csrftoken = $.cookie('csrftoken');
-	//url de la peticion
-	url = 'http://'+window.location.host+'/api/perdido/edit/'+id;
+	var r = confirm("¿Está seguro que quiere eliminar?");
+	if (r == true) {
+	    	
+		//obtiene el csrftoken
+		var csrftoken = $.cookie('csrftoken');
+		//url de la peticion
+		url = 'http://'+window.location.host+'/api/perdido/edit/'+id;
 
-	//configura el csrftoken a la peticion ajax
-	$.ajaxSetup({
-    	beforeSend: function(xhr, settings) {
-        xhr.setRequestHeader("X-CSRFToken", csrftoken);
-    	}
-	});
+		//configura el csrftoken a la peticion ajax
+		$.ajaxSetup({
+	    	beforeSend: function(xhr, settings) {
+	        xhr.setRequestHeader("X-CSRFToken", csrftoken);
+	    	}
+		});
 
-	//peticion ajax
-	$.ajax({
-		type: 'DELETE',
-		url: url
-	})
-	.done(function(response){
-		//$("#gretting").text(response)
-		console.log("perdido eliminado");
-		$("#id-perd-"+id).remove()
-	})
-	.fail(function(error){
-		console.log("fail")
-		console.log("no se pudo borrar el perdido")
-		
-	})
-	.always(function(){
-		//console.log("completo")
-		console.log("always");
-	});
+		//peticion ajax
+		$.ajax({
+			type: 'DELETE',
+			url: url
+		})
+		.done(function(response){
+			//$("#gretting").text(response)
+			console.log("perdido eliminado");
+			//$("#id-perd-"+id).remove()
+		})
+		.fail(function(error){
+			console.log("fail")
+			console.log("no se pudo borrar el perdido")
+			
+		})
+		.always(function(){
+			//console.log("completo")
+			console.log("always");
+		});
+
+	} else {
+	    
+	}
 }
